@@ -1,9 +1,20 @@
-const pools = require('express').Router();
-const poolsData = require('../data/pools');
-const status = require('../utils/statusMessages');
-const { validator } = require('../utils/validator');
-const { getWagerById } = require('../utils/pools');
-const {
+import { Router } from 'express';
+import {
+  acceptWager,
+  addUser,
+  addWager,
+  completeWager,
+  createPool,
+  deletePool,
+  getPoolById,
+  getUserPools,
+  removeWager,
+  transferWinnerPoints
+} from '../data/pools';
+import { badRequest, created, doesNotExist, serverError, success } from '../utils/statusMessages';
+import { validator } from '../utils/validator';
+import { getWagerById } from '../utils/pools';
+import {
   getUserPoolsQuery,
   defaultPoolParams,
   postPoolBody,
@@ -12,18 +23,20 @@ const {
   defaultWagerParams,
   patchWagerBody,
   completeWagerBody
-} = require('./validation/pools');
+} from './validation/pools';
+
+const pools = Router();
 
 pools.get('/', validator.query(getUserPoolsQuery), async (req, res) => {
   const { query: { pageNum, pageSize, userEmail } } = req;
-  const page = parseInt(pageNum) || 1;
-  const size = parseInt(pageSize) || 50;
+  const page = parseInt(pageNum as string) || 1;
+  const size = parseInt(pageSize as string) || 50;
 
-  const { items, totalItems, totalPages } = await poolsData.getUserPools(page, size, userEmail);
+  const { items, totalItems, totalPages } = await getUserPools(page, size, userEmail as string);
 
-  if (!items) return status.serverError(res, 'Failed', 'Failed to get pools');
+  if (!items) return serverError(res, 'Failed to get pools');
 
-  return status.success(res, {
+  return success(res, {
     items,
     pageNum: page,
     pageSize: size,
@@ -36,29 +49,33 @@ pools.get('/:poolId', validator.params(defaultPoolParams), async (req, res) => {
   const { params: { poolId } } = req;
 
   if (!poolId) {
-    return status.badRequest(res, { message: 'Invalid Pool ID provided' });
+    return badRequest(res, { message: 'Invalid Pool ID provided' });
   }
 
-  const pool = await poolsData.getPoolById(poolId);
+  const pool = await getPoolById(poolId as string);
 
-  return status.success(res, { ...pool });
+  if (!pool?._id) {
+    return doesNotExist(res, 'Pool');
+  }
+
+  return success(res, pool);
 });
 
 pools.post('/', validator.body(postPoolBody), async (req, res) => {
   const { body: { name, createdBy } } = req;
 
-  const pool = await poolsData.createPool(name, createdBy, [createdBy]);
+  const pool = await createPool(name, createdBy, [createdBy]);
 
-  return status.created(res, { ...pool });
+  return created(res, pool);
 });
 
 pools.delete('/:poolId',
   validator.params(defaultPoolParams), async (req, res) => {
     const { params: { poolId } } = req;
 
-    const deletedPool = await poolsData.deletePool(poolId);
+    const deletedPool = await deletePool(poolId);
 
-    return status.success(res, { deletedPool });
+    return success(res, { deletedPool });
   }
 );
 
@@ -67,9 +84,9 @@ pools.post('/:poolId/users',
   validator.body(postUserBody), async (req, res) => {
     const { params: { poolId }, body: { userEmail } } = req;
 
-    const addedUser = await poolsData.addUser(poolId, userEmail);
+    const addedUser = await addUser(poolId, userEmail);
 
-    return status.created(res, { ...addedUser });
+    return created(res, addedUser);
   }
 );
 
@@ -78,30 +95,37 @@ pools.post('/:poolId/wagers',
   validator.body(postWagerBody), async (req, res) => {
     const { params: { poolId }, body: { amount, description, createdBy, users } } = req;
 
-    const pool = await poolsData.getPoolById(poolId);
+    const pool = await getPoolById(poolId as string);
+
+    if (!pool._id) {
+      return doesNotExist(res, 'Pool');
+    }
 
     let hasError = false;
-    users.forEach(userEmail => {
+    users.forEach((userEmail: string) => {
       if (!pool.users.includes(userEmail)) {
         hasError = true;
       }
     });
 
     if (hasError) {
-      return status.badRequest(res, {
+      return badRequest(res, {
         message: 'User is not a member of this pool'
       });
     } else {
-      const addedWager = await poolsData.addWager(poolId, createdBy, {
+      const wager = {
         amount,
         description,
         users,
         isActive: false,
         isComplete: false,
-        activeUsers: [createdBy]
-      });
+        activeUsers: [createdBy],
+        createdBy,
+        winners: []
+      };
+      const addedWager = await addWager(poolId, createdBy, wager);
 
-      return status.created(res, { ...addedWager });
+      return created(res, addedWager);
     }
   }
 );
@@ -110,9 +134,13 @@ pools.delete('/:poolId/wagers/:wagerId',
   validator.params(defaultWagerParams), async (req, res) => {
     const { params: { poolId, wagerId } } = req;
 
-    const removedWager = await poolsData.removeWager(poolId, wagerId);
+    const removeResult = await removeWager(poolId, wagerId);
 
-    return status.success(res, { ...removedWager });
+    if ('error' in removeResult) {
+      return serverError(res, removeResult.error.message);
+    }
+
+    return success(res, { wagerId });
   }
 );
 
@@ -121,10 +149,15 @@ pools.patch('/:poolId/wagers/:wagerId/accept',
   validator.body(patchWagerBody), async (req, res) => {
     const { params: { poolId, wagerId }, body: { userEmail } } = req;
 
-    const pool = await poolsData.getPoolById(poolId);
+    const pool = await getPoolById(poolId);
+
+    if (!pool._id) {
+      return doesNotExist(res, 'Pool');
+    }
+
     const wager = getWagerById(pool, wagerId);
 
-    if (!!wager) {
+    if (wager._id) {
       const updatedWager = {
         ...wager,
         isActive: true,
@@ -134,12 +167,12 @@ pools.patch('/:poolId/wagers/:wagerId/accept',
         ]
       };
 
-      const updates = await poolsData.acceptWager(pool, updatedWager);
+      const updates = await acceptWager(pool, updatedWager);
 
-      return status.success(res, { updatedWager: updates });
+      return success(res, { updatedWager: updates });
     }
 
-    return status.doesNotExist(res, 'Wager', 'wagers', 'pool');
+    return doesNotExist(res, 'Wager', 'pool');
   }
 );
 
@@ -148,25 +181,25 @@ pools.patch('/:poolId/wagers/:wagerId/complete',
   validator.body(completeWagerBody), async (req, res) => {
     const { params: { poolId, wagerId }, body: { completedBy, winners } } = req;
 
-    const pool = await poolsData.getPoolById(poolId);
+    const pool = await getPoolById(poolId);
     const wager = getWagerById(pool, wagerId);
 
-    if (!!wager) {
+    if (wager._id) {
       const updatedWager = {
         ...wager,
         isComplete: true,
         winners
       };
 
-      const updates = await poolsData.completeWager(pool, updatedWager, completedBy);
+      const updates = await completeWager(pool, updatedWager, completedBy);
 
-      await poolsData.transferWinnerPoints(poolId, winners, wager.amount);
+      await transferWinnerPoints(poolId, winners, wager.amount);
 
-      return status.success(res, { updatedWager: updates });
+      return success(res, { updatedWager: updates });
     }
 
-    return status.doesNotExist(res, 'Wager', 'wagers', 'pool');
+    return doesNotExist(res, 'Wager', 'pool');
   }
 );
 
-module.exports = pools;
+export default pools;
