@@ -1,22 +1,24 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 require('dotenv').config();
-const express = require('express');
-const log = require('./utils/log');
-const { validationErrorHandler } = require('./utils/validator');
-const appInfo = require('../package.json');
+import express, { NextFunction, Request, Response } from 'express';
+import log from './utils/log';
+import api from './api';
+import { validationErrorHandler } from './utils/validator';
 
-const port = process.env.PORT || 5000;
+const appInfo = require('../package.json');
+const port = process.env.PORT ?? 5000;
 const app = express();
 
 app.use(express.json({ limit: '15mb'}));
 
-app.use((req, res, next) => {
+app.use((req: Request, res: Response, next: NextFunction) => {
   res.header('Access-Control-Allow-Origin', '*');
   res.header('Access-Control-Allow-Methods', 'GET, PUT, PATCH, POST, DELETE');
   res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
   next();
 });
 
-app.use('/api', require('./api'));
+app.use('/api', api);
 
 app.use(validationErrorHandler);
 
@@ -24,3 +26,5 @@ app.listen(port, () => {
   log.cool(`Pointless API v${appInfo.version}`);
   log.info(`Listening on port ${port}`);
 });
+
+export { appInfo }
